+++ packages/core/core/src/ConfigResolver.js
@@ -7,7 +7,7 @@ import type {
   PackageName
 } from '@parcel/types';
 import {resolveConfig} from '@parcel/utils';
-import Config from './Config';
+import Config from './ParcelConfig';
 import * as fs from '@parcel/fs';
 import {parse} from 'json5';
 import path from 'path';
