+++ packages/core/core/src/PackagerRunner.js
@@ -3,7 +3,7 @@
 import type {ParcelOptions, Blob, FilePath} from '@parcel/types';
 import type SourceMap from '@parcel/source-map';
 import type {Bundle as InternalBundle} from './types';
-import type Config from './Config';
+import type Config from './ParcelConfig';
 import type InternalBundleGraph from './BundleGraph';
 
 import {Readable} from 'stream';
