+++ packages/core/core/src/ReporterRunner.js
@@ -4,7 +4,7 @@ import type {ParcelOptions, ReporterEvent} from '@parcel/types';
 
 import {bundleToInternal, NamedBundle} from './public/Bundle';
 import {bus} from '@parcel/workers';
-import Config from './Config';
+import Config from './ParcelConfig';
 import logger from '@parcel/logger';
 import nullthrows from 'nullthrows';
